refactor(ios-tts-manager): split manager lifecycle, playback, and policy concerns
- replace the monolithic `iOS/KeyVox iOS/Core/TTS/TTSManager.swift` with a dedicated `iOS/KeyVox iOS/Core/TTS/TTSManager/` directory
- keep `TTSManager.swift` focused on the main type, shared published state, dependencies, callback wiring, and core initialization
- move app lifecycle, background handoff, background-task ownership, and idle-sleep updates into `TTSManager+AppLifecycle.swift`
- move request startup, clipboard and pending-request entry points, replay commands, and stop/resume playback commands into `TTSManager+Playback.swift`
- move playback completion, error handling, replay persistence, preparation progress, and state transitions into `TTSManager+State.swift`
- extract pure manager policy decisions into `TTSManagerPolicy.swift`, including idle-sleep prevention, active-state checks, preparation-view suppression, and background-task eligibility
- add `iOS/KeyVoxiOSTests/Core/TTS/TTSManager/TTSManagerPolicyTests.swift` to cover the extracted manager policy behavior
- update `iOS/KeyVox iOS.xcodeproj/project.pbxproj` to register the new `TTSManager` group and source files
- remove the old monolithic setter restriction on manager-owned published state so the split sibling files can still mutate the coordinator-backed manager state safely
